+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 === HEAD
 
+* Support use as a Rework plugin.
+* Change the input from a file path to a String.
+
 === 0.1.1 (November 21, 2013)
 
 * Simplify IE 8 opacity declaration.